fix spmat

+++ spmat.h
@@ -17,13 +17,13 @@ typedef struct _spmat {
 
 	/* Adds row i the matrix. Called before any other call,
 	 * exactly n times in order (i = 0 to n-1) */
-	void	(*add_row)(struct _spmat *A, const int *row, int i);
+	void	(*add_row)(struct _spmat *A, const double *row, int i);
 
 	/* Frees all resources used by A */
 	void	(*free)(struct _spmat *A);
 
 	/* Multiplies matrix A by vector v, into result (result is pre-allocated) */
-	void	(*mult)(const struct _spmat *A, const int *v, int *result);
+	void	(*mult)(const struct _spmat *A, const double *v, double *result);
 
 	/* Private field for inner implementation.
 	 * Should not be read or modified externally */
